feat: refactor models into packages, create deploy script after push, refactor and clean pom.xml
