+++ src/main/java/com/aerotrack/model/protocol/ScanQueryResponse.java
@@ -1,5 +1,6 @@
-package com.aerotrack.model;
+package com.aerotrack.model.protocol;
 
+import com.aerotrack.model.entities.FlightPair;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
